Fix crash type overrides (#18)
* Fix crash type overrides

ExceptionPostOptions and MinidumpPostOptions were implemented without defaults for ExceptionType and MinidumpType respectively. The thought was that they would get initialized to null, but it turns out they get initialized to 0 and thus checking if they are equal to null will never work. We need to update them to override if the value is 0 (Unknown) instead and set the default value to Unknown.

Fixes #17

* Updated tests